Collision Detection - cannot travel through walls
Improved the collision detection by modifying some parts such as how the
World Position is calculated for the blocks (offset + length/2) instead
of (x||y + offset + length/2). The additional X or Y was throwing the
value off giving the bizarre behaviours.

It still exists that the hero can move into the wall where there is a
gap between blocks but cannot pass through it into the wall / no go
zone.