Refactored Movement logic into Controllers
Created an abstract method that both AI and Hero Controller override.
This takes a BlockedDirections object as a parameter.

Now the ProcessInput function calls the process method on the
controllers and they handle the logic.

AI as been disabled while other effects are worked on (as I think AI can
be added last as a nice touch but not a necessity). The system may need
refactoring again as the AI controller controls two pawns but only gets
one BlockedDirections, maybe better off handling the blockeddirections
within the process() function as opposed to outside?